+++ src/includes/core/actions.php
@@ -54,6 +54,7 @@ add_action( 'wp_enqueue_scripts',       'bbp_enqueue_scripts',        10    );
 add_action( 'wp_head',                  'bbp_head',                   10    );
 add_action( 'wp_footer',                'bbp_footer',                 10    );
 add_action( 'transition_post_status',   'bbp_transition_post_status', 10, 3 );
+add_action( 'post_updated',             'bbp_post_updated',           10, 3 );
 
 /**
  * bbp_loaded - Attached to 'plugins_loaded' above
@@ -178,7 +179,7 @@ add_action( 'before_delete_post', 'bbp_delete_forum'  );
 // After Deleted/Trashed/Untrashed Forum
 add_action( 'trashed_post',   'bbp_trashed_forum'   );
 add_action( 'untrashed_post', 'bbp_untrashed_forum' );
-add_action( 'deleted_post',   'bbp_deleted_forum'   );
+add_action( 'deleted_post',   'bbp_deleted_forum', 10, 2 );
 
 // Auto trash/untrash/delete a forums topics
 add_action( 'bbp_delete_forum',  'bbp_delete_forum_topics',  10 );
@@ -272,8 +273,14 @@ add_action( 'bbp_unspammed_reply',  'bbp_update_reply_walker' );
 add_action( 'bbp_approved_reply',   'bbp_update_reply_walker' );
 add_action( 'bbp_unapproved_reply', 'bbp_update_reply_walker' );
 
-// Update counts when topics and replies are created or change status
+// Update counts from persisted post status changes
 add_action( 'bbp_transition_post_status', 'bbp_update_counts_on_transition_post_status', 10, 3 );
+add_action( 'bbp_transition_post_status', 'bbp_update_forum_subforum_count_on_transition_post_status', 10, 3 );
+
+// Update parent subforum counts after updates and permanent deletion
+add_action( 'bbp_deleted_forum',   'bbp_reparent_forum_subforums', 9, 2 );
+add_action( 'bbp_deleted_forum',   'bbp_update_parent_forum_subforum_count', 10, 2 );
+add_action( 'bbp_post_updated',    'bbp_update_forum_subforum_counts_on_post_updated', 10, 3 );
 
 // Update user topic & reply counts
 add_action( 'bbp_deleted_topic', 'bbp_decrease_user_topic_count' );